+++ src/YaNco.Abstractions/TypeMapping/IToAbapValueConverter.cs
@@ -1,6 +1,6 @@
 ﻿using LanguageExt;
 
-namespace Dbosoft.YaNco.Converters
+namespace Dbosoft.YaNco.TypeMapping
 {
     public interface IToAbapValueConverter<T>
     {
