+++ src/YaNco.Core/TypeMapping/CachingConverterResolver.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Dbosoft.YaNco.Converters
+namespace Dbosoft.YaNco.TypeMapping
 {
     public class CachingConverterResolver : IRfcConverterResolver
     {
