+++ src/YaNco.Core/TypeMapping/DefaultConverterResolver.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Dbosoft.YaNco.Converters
+namespace Dbosoft.YaNco.TypeMapping
 {
     public class DefaultConverterResolver : IRfcConverterResolver
     {
