+++ src/YaNco.Primitives/TypeMapping/AbapIntValue.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Dbosoft.YaNco.Converters
+namespace Dbosoft.YaNco.TypeMapping
 {
     public class AbapIntValue : AbapValue, IConvertible
     {
