+++ src/YaNco.Primitives/TypeMapping/AbapStringValue.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Globalization;
 
-namespace Dbosoft.YaNco.Converters
+namespace Dbosoft.YaNco.TypeMapping
 {
     public class AbapStringValue : AbapValue, IConvertible
     {
